feat: rework sidebar and layout CSS
