+++ frontend/js/src/explore/fresh-releases/ReleaseFilters.tsx
@@ -1,5 +1,10 @@
 import * as React from "react";
-import { faChevronDown, faChevronUp } from "@fortawesome/free-solid-svg-icons";
+import {
+  faChevronDown,
+  faChevronUp,
+  faGear,
+  faXmark,
+} from "@fortawesome/free-solid-svg-icons";
 import { faCircleXmark } from "@fortawesome/free-regular-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import Switch from "../../components/Switch";
@@ -47,6 +52,12 @@ export default function ReleaseFilters(props: ReleaseFiltersProps) {
     releaseCardGridRef,
     pageType,
   } = props;
+
+  // TODO: Refactor this functionality in a reusable Sidebar component
+  const [isSidebarOpen, setIsSidebarOpen] = React.useState<boolean>(false);
+  const toggleSidebar = () => {
+    setIsSidebarOpen(!isSidebarOpen);
+  };
 
   const [checkedList, setCheckedList] = React.useState<
     Array<string | undefined>
@@ -175,189 +186,200 @@ export default function ReleaseFilters(props: ReleaseFiltersProps) {
   ]);
 
   return (
-    <div className="sidebar settings-navbar">
-      <div id="sidebar-fresh-release">
-        <p>Fresh Releases</p>
-        <p>Listen to recent releases, and browse what&apos;s dropping soon.</p>
-        <p>
-          Check out all releases worldwide, or just from artists you&apos;ve
-          listened to before, with &apos;for you&apos;.
-        </p>
-      </div>
-      <div className="sidenav-content-grid">
-        <h4>
-          {filtersOpen ? (
-            <FontAwesomeIcon icon={faChevronDown} onClick={toggleFilters} />
-          ) : (
-            <FontAwesomeIcon icon={faChevronUp} onClick={toggleFilters} />
-          )}
-          {"  "}
-          <b>Filter</b>
-        </h4>
-
-        {filtersOpen && (
-          <>
-            <div id="range">Range: </div>
-
-            <div className="input-group">
-              <select
-                id="style"
-                className="form-control"
-                value={range}
-                onChange={handleRangeDropdown}
-              >
-                <option value={filterRangeOptions.week.key}>
-                  {filterRangeOptions.week.label}
-                </option>
-                <option value={filterRangeOptions.month.key}>
-                  {filterRangeOptions.month.label}
-                </option>
-                <option value={filterRangeOptions.three_months.key}>
-                  {filterRangeOptions.three_months.label}
-                </option>
-              </select>
-            </div>
-            <Switch
-              id="date-filter-item-past"
-              value="past"
-              checked={showPastReleases}
-              onChange={(e) => setShowPastReleases(!showPastReleases)}
-              switchLabel="Past"
-            />
-            <Switch
-              id="date-filter-item-future"
-              value="future"
-              checked={showFutureReleases}
-              onChange={(e) => setShowFutureReleases(!showFutureReleases)}
-              switchLabel="Future"
-            />
-            {allFilters.releaseTypes.length > 0 && (
-              <>
-                <span id="types">Types:</span>
-                {allFilters.releaseTypes?.map((type, index) => (
-                  <Switch
-                    id={`filters-item-${index}`}
-                    key={`filters-item-${type}`}
-                    value={type}
-                    checked={checkedList?.includes(type)}
-                    onChange={handleFilterChange}
-                    switchLabel={type}
-                  />
-                ))}
-              </>
+    <>
+      <div className={`sidebar ${isSidebarOpen ? "open" : ""}`}>
+        <div id="sidebar-fresh-release">
+          <p>Fresh Releases</p>
+          <p>
+            Listen to recent releases, and browse what&apos;s dropping soon.
+          </p>
+          <p>
+            Check out all releases worldwide, or just from artists you&apos;ve
+            listened to before, with &apos;for you&apos;.
+          </p>
+        </div>
+        <div className="sidenav-content-grid">
+          <h4>
+            {filtersOpen ? (
+              <FontAwesomeIcon icon={faChevronDown} onClick={toggleFilters} />
+            ) : (
+              <FontAwesomeIcon icon={faChevronUp} onClick={toggleFilters} />
             )}
+            {"  "}
+            <b>Filter</b>
+          </h4>
 
-            {allFilters.releaseTags.length > 0 && (
-              <>
-                <span id="tags">Include (only):</span>
-                <select
-                  id="include-tags"
-                  className="form-control"
-                  value=""
-                  onChange={handleIncludeTagChange}
-                >
-                  <option value="" disabled>
-                    genre/tag...
-                  </option>
-                  {allFilters.releaseTags
-                    ?.filter((tag) => !releaseTagsCheckList.includes(tag))
-                    ?.map((tag, index) => (
-                      <option value={tag} key={tag}>
-                        {tag}
-                      </option>
-                    ))}
-                </select>
+          {filtersOpen && (
+            <>
+              <div id="range">Range: </div>
 
-                <div className="release-tags">
-                  {releaseTagsCheckList?.map((tag, index) => (
-                    <div id={`include-tag-item-${index}`} className="tags">
-                      <span className="release-tag-name">{tag}</span>
-                      <FontAwesomeIcon
-                        icon={faCircleXmark}
-                        onClick={() => removeFilterTag(tag!)}
-                      />
-                    </div>
-                  ))}
-                </div>
-
-                <span id="tags">Exclude:</span>
+              <div className="input-group">
                 <select
                   id="style"
                   className="form-control"
-                  value=""
-                  onChange={handleExcludeTagChange}
+                  value={range}
+                  onChange={handleRangeDropdown}
                 >
-                  <option value="" disabled>
-                    genre/tag...
+                  <option value={filterRangeOptions.week.key}>
+                    {filterRangeOptions.week.label}
                   </option>
-                  {allFilters.releaseTags
-                    ?.filter(
-                      (tag) => !releaseTagsExcludeCheckList.includes(tag)
-                    )
-                    ?.map((tag, index) => (
-                      <option value={tag} key={tag}>
-                        {tag}
-                      </option>
-                    ))}
+                  <option value={filterRangeOptions.month.key}>
+                    {filterRangeOptions.month.label}
+                  </option>
+                  <option value={filterRangeOptions.three_months.key}>
+                    {filterRangeOptions.three_months.label}
+                  </option>
                 </select>
-
-                <div className="release-tags">
-                  {releaseTagsExcludeCheckList?.map((tag, index) => (
-                    <div id={`exclude-tag-item-${index}`} className="tags">
-                      <span className="release-tag-name">{tag}</span>
-                      <FontAwesomeIcon
-                        icon={faCircleXmark}
-                        onClick={() => removeExcludeTag(tag!)}
-                      />
-                    </div>
+              </div>
+              <Switch
+                id="date-filter-item-past"
+                value="past"
+                checked={showPastReleases}
+                onChange={(e) => setShowPastReleases(!showPastReleases)}
+                switchLabel="Past"
+              />
+              <Switch
+                id="date-filter-item-future"
+                value="future"
+                checked={showFutureReleases}
+                onChange={(e) => setShowFutureReleases(!showFutureReleases)}
+                switchLabel="Future"
+              />
+              {allFilters.releaseTypes.length > 0 && (
+                <>
+                  <span id="types">Types:</span>
+                  {allFilters.releaseTypes?.map((type, index) => (
+                    <Switch
+                      id={`filters-item-${index}`}
+                      key={`filters-item-${type}`}
+                      value={type}
+                      checked={checkedList?.includes(type)}
+                      onChange={handleFilterChange}
+                      switchLabel={type}
+                    />
                   ))}
-                </div>
-              </>
-            )}
-          </>
-        )}
-      </div>
-      <div className="sidenav-content-grid">
-        <h4>
-          {displayOpen ? (
-            <FontAwesomeIcon icon={faChevronDown} onClick={toggleDisplay} />
-          ) : (
-            <FontAwesomeIcon icon={faChevronUp} onClick={toggleDisplay} />
+                </>
+              )}
+
+              {allFilters.releaseTags.length > 0 && (
+                <>
+                  <span id="tags">Include (only):</span>
+                  <select
+                    id="include-tags"
+                    className="form-control"
+                    value=""
+                    onChange={handleIncludeTagChange}
+                  >
+                    <option value="" disabled>
+                      genre/tag...
+                    </option>
+                    {allFilters.releaseTags
+                      ?.filter((tag) => !releaseTagsCheckList.includes(tag))
+                      ?.map((tag, index) => (
+                        <option value={tag} key={tag}>
+                          {tag}
+                        </option>
+                      ))}
+                  </select>
+
+                  <div className="release-tags">
+                    {releaseTagsCheckList?.map((tag, index) => (
+                      <div id={`include-tag-item-${index}`} className="tags">
+                        <span className="release-tag-name">{tag}</span>
+                        <FontAwesomeIcon
+                          icon={faCircleXmark}
+                          onClick={() => removeFilterTag(tag!)}
+                        />
+                      </div>
+                    ))}
+                  </div>
+
+                  <span id="tags">Exclude:</span>
+                  <select
+                    id="style"
+                    className="form-control"
+                    value=""
+                    onChange={handleExcludeTagChange}
+                  >
+                    <option value="" disabled>
+                      genre/tag...
+                    </option>
+                    {allFilters.releaseTags
+                      ?.filter(
+                        (tag) => !releaseTagsExcludeCheckList.includes(tag)
+                      )
+                      ?.map((tag, index) => (
+                        <option value={tag} key={tag}>
+                          {tag}
+                        </option>
+                      ))}
+                  </select>
+
+                  <div className="release-tags">
+                    {releaseTagsExcludeCheckList?.map((tag, index) => (
+                      <div id={`exclude-tag-item-${index}`} className="tags">
+                        <span className="release-tag-name">{tag}</span>
+                        <FontAwesomeIcon
+                          icon={faCircleXmark}
+                          onClick={() => removeExcludeTag(tag!)}
+                        />
+                      </div>
+                    ))}
+                  </div>
+                </>
+              )}
+            </>
           )}
-          {"  "}
-          <b>Display</b>
-        </h4>
-
-        {displayOpen && (
-          <>
-            <Switch
-              id="coverart-only"
-              value="coverart-only"
-              checked={coverartOnly}
-              onChange={(e) => setCoverartOnly(!coverartOnly)}
-              switchLabel="Only Releases with artwork"
-            />
-
-            {Object.keys(displaySettings).map((setting, index) =>
-              (setting === "Tags" && allFilters.releaseTags.length === 0) ||
-              (setting === "Listens" && !totalListenCount) ? null : (
-                <Switch
-                  id={`display-item-${index}`}
-                  key={`display-item-${setting}`}
-                  value={setting}
-                  checked={
-                    displaySettings[setting as DisplaySettingsPropertiesEnum]
-                  }
-                  onChange={(e) =>
-                    toggleSettings(setting as DisplaySettingsPropertiesEnum)
-                  }
-                  switchLabel={setting}
-                />
-              )
+        </div>
+        <div className="sidenav-content-grid">
+          <h4>
+            {displayOpen ? (
+              <FontAwesomeIcon icon={faChevronDown} onClick={toggleDisplay} />
+            ) : (
+              <FontAwesomeIcon icon={faChevronUp} onClick={toggleDisplay} />
             )}
-          </>
-        )}
+            {"  "}
+            <b>Display</b>
+          </h4>
+
+          {displayOpen && (
+            <>
+              <Switch
+                id="coverart-only"
+                value="coverart-only"
+                checked={coverartOnly}
+                onChange={(e) => setCoverartOnly(!coverartOnly)}
+                switchLabel="Only Releases with artwork"
+              />
+
+              {Object.keys(displaySettings).map((setting, index) =>
+                (setting === "Tags" && allFilters.releaseTags.length === 0) ||
+                (setting === "Listens" && !totalListenCount) ? null : (
+                  <Switch
+                    id={`display-item-${index}`}
+                    key={`display-item-${setting}`}
+                    value={setting}
+                    checked={
+                      displaySettings[setting as DisplaySettingsPropertiesEnum]
+                    }
+                    onChange={(e) =>
+                      toggleSettings(setting as DisplaySettingsPropertiesEnum)
+                    }
+                    switchLabel={setting}
+                  />
+                )
+              )}
+            </>
+          )}
+        </div>
       </div>
-    </div>
+      <button
+        className={`toggle-sidebar-button ${isSidebarOpen ? "open" : ""}`}
+        onClick={toggleSidebar}
+        type="button"
+      >
+        <FontAwesomeIcon icon={isSidebarOpen ? faXmark : faGear} size="2x" />
+      </button>
+    </>
   );
 }
